requirements.txt
asgiref==3.9.2
crispy-bootstrap5==2025.6
Django==6.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
Markdown==3.10
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
xlsxwriter==3.2.9
